+++ src/clients/default.ts
@@ -1,23 +1,24 @@
 import { Utils, Wallet } from 'xpring-common-js'
-
 import bigInt, { BigInteger } from 'big-integer'
-
-import { RawTransactionStatus } from '../raw-transaction'
-import LegacyGRPCNetworkClientNode from './legacy-grpc-network-client'
-import LegacyGRPCNetworkClientWeb from './legacy-grpc-network-client.web'
-import { XpringClientDecorator } from '../xpring-client-decorator'
-import TransactionStatus from '../transaction-status'
+import { XpringClientDecorator } from '../utils/xpring-client-decorator'
+import TransactionStatus from '../utils/transaction-status'
+import RawTransaction from '../utils/raw-transaction'
+import GRPCNetworkClientNode from '../grpc/node'
+import GRPCNetworkClientWeb from '../grpc/web'
+import { XRPDropsAmount } from '../generated/web/rpc/v1/amount_pb' // TODO would be nice to have 0 references to generated files here
 import isNode from '../utils'
-import { LegacyNetworkClient } from './legacy-network-client'
-import XpringClientErrorMessages from '../xpring-client-error-messages'
+import { NetworkClientDecorator } from '../utils/network-client-decorator'
+import XpringClientErrorMessages from '../utils/xpring-client-error-messages'
 
-/** A margin to pad the current ledger sequence with when submitting transactions. */
-const ledgerSequenceMargin = 10
+// TODO(keefertaylor): Re-enable this rule when this class is fully implemented.
+/* eslint-disable @typescript-eslint/require-await */
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/* eslint-disable class-methods-use-this */
 
 /**
  * DefaultXpringClient is a client which interacts with the Xpring platform.
  */
-class LegacyDefaultXpringClient implements XpringClientDecorator {
+class DefaultXpringClient implements XpringClientDecorator {
   /**
    * Create a new DefaultXpringClient.
    *
@@ -29,10 +30,10 @@ class LegacyDefaultXpringClient implements XpringClientDecorator {
   public static defaultXpringClientWithEndpoint(
     grpcURL: string,
     forceWeb = false,
-  ): LegacyDefaultXpringClient {
+  ): DefaultXpringClient {
     return isNode() && !forceWeb
-      ? new LegacyDefaultXpringClient(new LegacyGRPCNetworkClientNode(grpcURL))
-      : new LegacyDefaultXpringClient(new LegacyGRPCNetworkClientWeb(grpcURL))
+      ? new DefaultXpringClient(new GRPCNetworkClientNode(grpcURL))
+      : new DefaultXpringClient(new GRPCNetworkClientWeb(grpcURL))
   }
 
   /**
@@ -42,7 +43,7 @@ class LegacyDefaultXpringClient implements XpringClientDecorator {
    *
    * @param networkClient A network client which will manage remote RPCs to Rippled.
    */
-  public constructor(private readonly networkClient: LegacyNetworkClient) {}
+  public constructor(private readonly networkClient: NetworkClientDecorator) {}
 
   /**
    * Retrieve the balance for the given address.
@@ -51,24 +52,26 @@ class LegacyDefaultXpringClient implements XpringClientDecorator {
    * @returns A `BigInteger` representing the number of drops of XRP in the account.
    */
   public async getBalance(address: string): Promise<BigInteger> {
-    if (!Utils.isValidXAddress(address)) {
+    const classicAddress = Utils.decodeXAddress(address)
+    if (!classicAddress) {
       return Promise.reject(
         new Error(XpringClientErrorMessages.xAddressRequired),
       )
     }
 
-    return this.networkClient
-      .getAccountInfo(address)
-      .then(async (accountInfo) => {
-        const balance = accountInfo.getBalance()
-        if (balance === undefined) {
-          return Promise.reject(
-            new Error(XpringClientErrorMessages.malformedResponse),
-          )
-        }
+    const accountInfo = await this.networkClient.getAccountInfo(
+      classicAddress.address,
+    )
+    const accountData = accountInfo.getAccountData()
+    if (!accountData) {
+      throw new Error(XpringClientErrorMessages.malformedResponse)
+    }
 
-        return bigInt(balance.getDrops())
-      })
+    const balance = accountData.getBalance()
+    if (!balance) {
+      throw new Error(XpringClientErrorMessages.malformedResponse)
+    }
+    return bigInt(balance.getDrops())
   }
 
   /**
@@ -103,43 +106,42 @@ class LegacyDefaultXpringClient implements XpringClientDecorator {
    * @returns A promise which resolves to a string representing the hash of the submitted transaction.
    */
   public async send(
-    amount: BigInteger | number | string,
-    destination: string,
-    sender: Wallet,
+    _amount: BigInteger | number | string,
+    _destination: string,
+    _sender: Wallet,
   ): Promise<string> {
-    if (!Utils.isValidXAddress(destination)) {
-      throw new Error(XpringClientErrorMessages.xAddressRequired)
-    }
-    const signedTransaction = await this.networkClient.createSignedTransaction(
-      amount.toString(),
-      destination,
-      sender,
-      ledgerSequenceMargin,
-    )
-    const response = await this.networkClient.submitSignedTransaction(
-      signedTransaction,
-    )
-
-    const transactionBlob = response.getTransactionBlob()
-    const transactionHash = Utils.transactionBlobToTransactionHash(
-      transactionBlob,
-    )
-    if (!transactionHash) {
-      throw new Error(XpringClientErrorMessages.malformedResponse)
-    }
-    return transactionHash
+    throw new Error(XpringClientErrorMessages.unimplemented)
   }
 
   public async getLastValidatedLedgerSequence(): Promise<number> {
-    const ledgerSequence = await this.networkClient.getLatestValidatedLedgerSequence()
-    return ledgerSequence.getIndex()
+    const getFeeResponse = await this.networkClient.getFee()
+    return getFeeResponse.getLedgerCurrentIndex()
   }
 
   public async getRawTransactionStatus(
     transactionHash: string,
-  ): Promise<RawTransactionStatus> {
-    return this.networkClient.getTransactionStatus(transactionHash)
+  ): Promise<RawTransaction> {
+    const getTxResponse = await this.networkClient.getTx(transactionHash)
+    return new RawTransaction(getTxResponse)
+  }
+
+  // TODO Keefer implement method and remove tslint ignore and fix tsconfig nounusedlocals
+  // tslint:disable-next-line
+  private async getMinimumFee(): Promise<XRPDropsAmount> {
+    const getFeeResponse = await this.networkClient.getFee()
+
+    const fee = getFeeResponse.getDrops()
+    if (!fee) {
+      throw new Error(XpringClientErrorMessages.malformedResponse)
+    }
+
+    const minimumFee = fee.getMinimumFee()
+    if (!minimumFee) {
+      throw new Error(XpringClientErrorMessages.malformedResponse)
+    }
+
+    return minimumFee
   }
 }
 
-export default LegacyDefaultXpringClient
+export default DefaultXpringClient
