+++ src/grpc/web.ts
@@ -2,23 +2,23 @@ import { Utils } from 'xpring-common-js'
 import {
   GetAccountInfoRequest,
   GetAccountInfoResponse,
-} from './generated/web/rpc/v1/account_info_pb'
-import { GetFeeRequest, GetFeeResponse } from './generated/web/rpc/v1/fee_pb'
-import { GetTxRequest, GetTxResponse } from './generated/web/rpc/v1/tx_pb'
+} from '../generated/web/rpc/v1/account_info_pb'
+import { GetFeeRequest, GetFeeResponse } from '../generated/web/rpc/v1/fee_pb'
+import { GetTxRequest, GetTxResponse } from '../generated/web/rpc/v1/tx_pb'
 import {
   SubmitTransactionRequest,
   SubmitTransactionResponse,
-} from './generated/web/rpc/v1/submit_pb'
-import { XRPLedgerAPIServiceClient } from './generated/web/rpc/v1/xrp_ledger_grpc_web_pb'
+} from '../generated/web/rpc/v1/submit_pb'
+import { XRPLedgerAPIServiceClient } from '../generated/web/rpc/v1/xrp_ledger_grpc_web_pb'
 
-import { AccountAddress } from './generated/web/rpc/v1/amount_pb'
-import isNode from './utils'
-import { NetworkClient } from './network-client'
+import { AccountAddress } from '../generated/web/rpc/v1/amount_pb'
+import isNode from '../utils'
+import { NetworkClientDecorator } from '../utils/network-client-decorator'
 
 /**
  * A GRPC Based network client.
  */
-class GRPCNetworkClientWeb implements NetworkClient {
+class GRPCNetworkClientWeb implements NetworkClientDecorator {
   private readonly grpcClient: XRPLedgerAPIServiceClient
 
   public constructor(grpcURL: string) {
