+++ src/legacy/grpc/node.ts
@@ -1,41 +1,34 @@
-import { Signer, Wallet } from 'xpring-common-js'
-import { XRPLedgerAPIClient } from '../generated/web/legacy/xrp_ledger_grpc_web_pb'
-import { AccountInfo } from '../generated/web/legacy/account_info_pb'
-import { Fee } from '../generated/web/legacy/fee_pb'
-import { GetFeeRequest } from '../generated/web/legacy/get_fee_request_pb'
-import { GetAccountInfoRequest } from '../generated/web/legacy/get_account_info_request_pb'
-import { SubmitSignedTransactionRequest } from '../generated/web/legacy/submit_signed_transaction_request_pb'
-import { SubmitSignedTransactionResponse } from '../generated/web/legacy/submit_signed_transaction_response_pb'
-import { GetLatestValidatedLedgerSequenceRequest } from '../generated/web/legacy/get_latest_validated_ledger_sequence_request_pb'
-import { LedgerSequence } from '../generated/web/legacy/ledger_sequence_pb'
-import { GetTransactionStatusRequest } from '../generated/web/legacy/get_transaction_status_request_pb'
-import { TransactionStatus } from '../generated/web/legacy/transaction_status_pb'
-import { XRPAmount } from '../generated/node/legacy/xrp_amount_pb'
-import { Payment } from '../generated/node/legacy/payment_pb'
-import { Transaction } from '../generated/node/legacy/transaction_pb'
-import isNode from '../utils'
-import { SignedTransaction } from '../generated/web/legacy/signed_transaction_pb'
-import { LegacyNetworkClient } from './legacy-network-client'
-import XpringClientErrorMessages from '../xpring-client-error-messages'
+import { credentials } from 'grpc'
+import { Wallet, Signer } from 'xpring-common-js'
+import { XRPLedgerAPIClient } from '../../generated/node/legacy/xrp_ledger_grpc_pb'
+import { AccountInfo } from '../../generated/node/legacy/account_info_pb'
+import { Fee } from '../../generated/node/legacy/fee_pb'
+import { GetFeeRequest } from '../../generated/node/legacy/get_fee_request_pb'
+import { GetAccountInfoRequest } from '../../generated/node/legacy/get_account_info_request_pb'
+import { SubmitSignedTransactionRequest } from '../../generated/node/legacy/submit_signed_transaction_request_pb'
+import { SubmitSignedTransactionResponse } from '../../generated/node/legacy/submit_signed_transaction_response_pb'
+import { GetLatestValidatedLedgerSequenceRequest } from '../../generated/node/legacy/get_latest_validated_ledger_sequence_request_pb'
+import { LedgerSequence } from '../../generated/node/legacy/ledger_sequence_pb'
+import { GetTransactionStatusRequest } from '../../generated/node/legacy/get_transaction_status_request_pb'
+import { TransactionStatus } from '../../generated/node/legacy/transaction_status_pb'
+import { XRPAmount } from '../../generated/node/legacy/xrp_amount_pb'
+import { Payment } from '../../generated/node/legacy/payment_pb'
+import { Transaction } from '../../generated/node/legacy/transaction_pb'
+import { SignedTransaction } from '../../generated/node/legacy/signed_transaction_pb'
+import { LegacyNetworkClient } from './network-client'
+import XpringClientErrorMessages from '../../utils/xpring-client-error-messages'
 
 /**
  * A GRPC Based network client.
  */
-class LegacyGRPCNetworkClientWeb implements LegacyNetworkClient {
+class LegacyGRPCNetworkClientNode implements LegacyNetworkClient {
   private readonly grpcClient: XRPLedgerAPIClient
 
   public constructor(grpcURL: string) {
-    if (isNode()) {
-      try {
-        // This polyfill hack enables XMLHttpRequest on the global node.js state
-        // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-        // @ts-ignore eslint-disable-line
-        global.XMLHttpRequest = require('xhr2') // eslint-disable-line
-      } catch {
-        // Swallow the error here for browsers
-      }
-    }
-    this.grpcClient = new XRPLedgerAPIClient(grpcURL)
+    this.grpcClient = new XRPLedgerAPIClient(
+      grpcURL,
+      credentials.createInsecure(),
+    )
   }
 
   public async getAccountInfo(address: string): Promise<AccountInfo> {
@@ -44,7 +37,6 @@ class LegacyGRPCNetworkClientWeb implements LegacyNetworkClient {
       getAccountInfoRequest.setAddress(address)
       this.grpcClient.getAccountInfo(
         getAccountInfoRequest,
-        {},
         (error, response): void => {
           if (error !== null || response === null) {
             reject(error)
@@ -59,7 +51,7 @@ class LegacyGRPCNetworkClientWeb implements LegacyNetworkClient {
   public async getFee(): Promise<Fee> {
     return new Promise((resolve, reject): void => {
       const getFeeRequest = new GetFeeRequest()
-      this.grpcClient.getFee(getFeeRequest, {}, (error, response): void => {
+      this.grpcClient.getFee(getFeeRequest, (error, response): void => {
         if (error !== null || response === null) {
           reject(error)
           return
@@ -77,7 +69,6 @@ class LegacyGRPCNetworkClientWeb implements LegacyNetworkClient {
       submitSignedTransactionRequest.setSignedTransaction(signedTransaction)
       this.grpcClient.submitSignedTransaction(
         submitSignedTransactionRequest,
-        {},
         (error, response): void => {
           if (error !== null || response === null) {
             reject(error)
@@ -94,7 +85,6 @@ class LegacyGRPCNetworkClientWeb implements LegacyNetworkClient {
       const getLatestValidatedLedgerSequenceRequest = new GetLatestValidatedLedgerSequenceRequest()
       this.grpcClient.getLatestValidatedLedgerSequence(
         getLatestValidatedLedgerSequenceRequest,
-        {},
         (error, response): void => {
           if (error !== null || response === null) {
             reject(error)
@@ -114,7 +104,6 @@ class LegacyGRPCNetworkClientWeb implements LegacyNetworkClient {
       getTransactionStatusRequest.setTransactionHash(transactionHash)
       this.grpcClient.getTransactionStatus(
         getTransactionStatusRequest,
-        {},
         (error, response): void => {
           if (error !== null || response === null) {
             reject(error)
@@ -167,4 +156,4 @@ class LegacyGRPCNetworkClientWeb implements LegacyNetworkClient {
   }
 }
 
-export default LegacyGRPCNetworkClientWeb
+export default LegacyGRPCNetworkClientNode
